+++ memoriter-website-gatsby/src/components/layout/cookie-banner/cookie-settings.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useState } from 'react';
 import { Link } from 'gatsby';
-import '../../../styles/wrapper/cookie-banner/cookie-settings.scss';
+import '../../../styles/layout/cookie-banner/cookie-settings.scss';
 import cookies from '../../../utils/cookies';
 
 interface props { //type definitions for props
