+++ memoriter-website-gatsby/src/styles/layout/cookie-banner/cookie-settings.scss
